+++ CHANGELOG.md
@@ -9,6 +9,17 @@ This is a high level and scarce summary of the changes between releases.  We
 would recommend to consult log of the [DataLad git
 repository](http://github.com/datalad/datalad-container) for more details.
 
+## 0.5.0 (Jul 12, 2019) -- damn-you-malicious-users
+
+### New features
+
+- The default result renderer for `containers-list` is now a custom
+  renderer that includes the container name in the output.
+
+### Fixes
+
+- Temporarily skip two tests relying on SingularityHub -- it is down.
+
 ## 0.4.0 (May 29, 2019) -- run-baby-run
 
 The minimum required DataLad version is now 0.11.5.